Conflicts:
	recipes/enigma2/enigma2.bb
	recipes/linux/linux-etxx00_3.1.0.bb
	recipes/openpli/openpli-bootlogo/bootlogo-hd.mvi
	recipes/openpli/openpli-bootlogo/bootlogo.jpg
	recipes/openpli/openpli-bootlogo/bootlogo.mvi
	recipes/openpli/openpli-bootlogo/switchoff.mvi

+++ recipes/linux/linux-etxx00_3.1.0.bb
@@ -1,7 +1,7 @@
 DESCRIPTION = "Linux kernel for ${MACHINE}"
 LICENSE = "GPL"
 
-MACHINE_KERNEL_PR_append = ".6"
+MACHINE_KERNEL_PR_append = ".12"
 
 DEPENDS = "mtd-minimal-nand-utils"
 RDEPENDS_kernel-image = "mtd-minimal-nand-utils"
@@ -17,6 +17,9 @@ SRC_URI += "http://www.et-view.com/download/linux-${PV}.tar.gz \
 	file://cxd2820r-enable-LNA-for-DVB-T.patch \
 	file://cxd2820r-changed-condition-to-break-out-from-wait-lock-loop.patch \
 	file://dvb-usb-smsdvb_fix_frontend.patch \
+	file://dvb-usb-rtl2832.patch \
+	file://cxd2820r-output-full-range-SNR.patch \
+	file://xc3028-fix-center-frequency.patch \
 	"
 
 SRC_URI_append_et5x00 = " file://disable_early_fb.patch"
@@ -45,12 +48,12 @@ kernel_do_install_append() {
 }
 
 pkg_postinst_kernel-image () {
-	if [ -f /${KERNEL_IMAGEDEST}/${KERNEL_IMAGETYPE}.gz ] ; then
-		if [ -d /proc/stb ] ; then
+	if [ "x$D" == "x" ]; then
+		if [ -f /${KERNEL_IMAGEDEST}/${KERNEL_IMAGETYPE}.gz ] ; then
 			flash_eraseall -j /dev/mtd1
 			nandwrite -p /dev/mtd1 /${KERNEL_IMAGEDEST}/${KERNEL_IMAGETYPE}.gz
+			rm -f /${KERNEL_IMAGEDEST}/${KERNEL_IMAGETYPE}.gz
 		fi
-		rm -f /${KERNEL_IMAGEDEST}/${KERNEL_IMAGETYPE}.gz
 	fi
 	true
 }
